Fix cursor overflow state.
Terminals act a little funny when you print past the end of the screen.  The cursor remains on the last column, but it will insert a newline if you try to print more text.

Before this patch we got this behavior wrong, in that we'd insert the newline immediately after the cursor overflowed, but before new text was inserted.  This screws up the case where a cursor overflow is followed by a cursor movement operation.

Fixing this exposed a major problem with row invalidation.  The way it was written, it would trust the rowNode index of whichever node it found in the DOM was the rowIndex that should be there.  In the case where rows have been renumbered, this may not be true.  The only way to invalidate is to to a full redraw of the visible rows (though we can skip some of the overhead of a full redraw_.)

This CL also adds a binary step command, 'bstep' to vtscope.py.

It also removes the 'device status' escape code from vttest-01.log, since the terminal's response would get displayed by `nc`.

BUG=chromium-os:24772
TEST=test_harness.html, 52/52 tests passed.

Change-Id: I2c496e0231767517468ba82df6dbd671add05cf4
Reviewed-on: https://gerrit.chromium.org/gerrit/13809
Reviewed-by: Zelidrag Hornung <zelidrag@chromium.org>
Tested-by: Robert Ginda <rginda@chromium.org>
Commit-Ready: Robert Ginda <rginda@chromium.org>