hterm: Add packaging script.
This adds bin/package.sh, which makes it easier to maintain the dev and stable "channels".
The dev channel is packaged as a crx which can be distributed outside the app store, while the stable package is a zip which can be uploaded to the chrome app store.

It also adds changelogs for both channels.
BUG=none
TEST=none

Change-Id: I519c23022343bca56c2111e7a48403d14be54079
Reviewed-on: https://gerrit.chromium.org/gerrit/16366
Commit-Ready: Robert Ginda <rginda@chromium.org>
Reviewed-by: Robert Ginda <rginda@chromium.org>
Tested-by: Robert Ginda <rginda@chromium.org>