hterm: Fix extra-long line wrapping (version 2).
We regressed a few versions ago such that inserting a line of text
that is more that twice the width of the terminal clips everything
after the first line wrap.  This happened back when we fixed line
wrapping to not wrap existing text.

This fix teaches Terminal..print() to wrap long lines, which should
be faster anyway.

This also adds a new testcase to catch future regressions.

There is a small addition change to add lib.UTF8Decoder to the list
of runtime deps for hterm_vt.js.

BUG=None
TEST=test_harness.html, 64/64 tests passed.

Change-Id: I4adf476cc8878ecfa276cf5db9266ce874cc3533
Reviewed-on: https://gerrit.chromium.org/gerrit/29388
Reviewed-by: David Benjamin <davidben@chromium.org>
Commit-Ready: Robert Ginda <rginda@chromium.org>
Tested-by: Robert Ginda <rginda@chromium.org>