hterm: Don't clear selection on copy
Clearing the selection on a non-keyboard copy was tricky, because
we don't want to do it right away, but the delay sucks if you're
trying to work quickly.

Unfortunately we *have* to clear the selection as part of the copy
process, so this patch restores it right after the copy completes.

BUG=none
TEST=text_harness.html, 64/64 tests passed.

Change-Id: I489093dcbebb170c92d64a6f4e11ef4f86fd97b1
Reviewed-on: https://gerrit.chromium.org/gerrit/31061
Reviewed-by: Marius Schilder <mschilder@google.com>
Commit-Ready: Robert Ginda <rginda@chromium.org>
Reviewed-by: Robert Ginda <rginda@chromium.org>
Tested-by: Robert Ginda <rginda@chromium.org>