saltpig: Initial commit of nacl collection v2 app
This is an experimental app that hosts a few binaries from naclports,
exposing their functionality over wam.FileSystem.  It's currently
hardcoded to only speak to the "Wash (tot)" and "Wash (dev)"
extensions.

You'll need to run the ./bin/get_nacl.sh file to fetch the necessary
nacl binaries from the web, and ./bin/mkdeps.sh to build the js
dependencies.

Change-Id: I5db5003903854bdc45c7c6b81eb482d2a47e7a26
Reviewed-on: https://chromium-review.googlesource.com/198908
Reviewed-by: Robert Ginda <rginda@chromium.org>
Tested-by: Robert Ginda <rginda@chromium.org>