hterm: Don't request Notification permission.
Notification.requestPermission is not honored (by Chrome, at least) if
it is not the direct result of a user action, so don't call it. Instead,
log a warning if the pref is turned on but notifications will still not
be displayed. (I added a note about rginda's proposed workaround, as
well.)

Change-Id: I96bd32d440c74d7977cb37b7d0bde357d42ab98b
Reviewed-on: https://chromium-review.googlesource.com/205539
Reviewed-by: Robert Ginda <rginda@chromium.org>
Tested-by: Robert Ginda <rginda@chromium.org>