nassh: Don't reload while windows are open.
Handle the chrome.runtime.onUpdateAvailable event so that we're
not automatically reloaded when there's a pending update.

When we notice an update available we show a unicode up-arrow in
each terminal window.  When all windows are closed we'll apply the
update.  Users can dismiss the update notification by clicking on
it.

Change-Id: I943abd63709fc08d0d2c702a4b3a27a2b5cec5c7
Reviewed-on: https://chromium-review.googlesource.com/210538
Reviewed-by: Antony Sargent <asargent@chromium.org>
Tested-by: Robert Ginda <rginda@chromium.org>