nassh: Add the axiom filesystem library and wash executables.
We now create an instance of an Axiom JsFileSystem on the
background page, and have a new "/html/hterm.html" page which
will run an interactive wash session over the file system.

We have the stock wash executables, and one called "hterm"
which can be used to create a new terminal window, and
one called "nassh" which starts a new ssh session.

Change-Id: I9df593bb2562127fc24f71d6ced25f2456bccc2b
Reviewed-on: https://chromium-review.googlesource.com/252401
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Rob Ginda <rginda@chromium.org>