nassh: Add the axiom filesystem library and wash executables.
We now create an instance of an Axiom JsFileSystem on the
background page, and have a new "/html/hterm.html" page which
will run an interactive wash session over the file system.

We have the stock wash executables, and one called "hterm"
which can be used to create a new terminal window, and
one called "nassh" which starts a new ssh session.

Change-Id: I4300569e8b4775a45753bf0c03e472a19c44a576
Reviewed-on: https://chromium-review.googlesource.com/252401
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Rob Ginda <rginda@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/257050