Update nassh after changes to ssh_client NaCl module
In ssh_client, now stdin/stdout/stderr always call OpenFile, and expect
to be returned whether the stream is a TTY. This change unifies "normal"
JavaScript streams and stdin/stdout/stderr, which used to be treated
specially.

Because /dev/tty and stdin are no longer shared (they can be different
when piping stdin, e.g. running scp). All input from the terminal must
be buffered instead of being sent directly to the plugin on the 0 fd.

This changes file input from a push to a pull model (w.r.t. the NaCl
module), which means that select/poll-ing will no longer work. As a result, the
JavaScript code pushes "OnReadReady" events to the NaCl module.

Change-Id: If6190f7ea8253928cc933b26d86e9f2a02ce6d94
Reviewed-on: https://chromium-review.googlesource.com/257080
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Rob Ginda <rginda@chromium.org>