ssh_client: Split stdin/stdout/stderr and /dev/tty
The previous code assumes that reading /dev/tty maps to stdin, and
writing to /dev/tty maps to stdout. This is not true when piping
stdin/stdout, which is necessary to support scp, for example.

This change primarily modifies the JsFile interface to require that
reading is a pulled rather than pushed. That is, the NaCl module must
request data to be read, which will be buffered on the JavaScript side.
This way, reading from stdin and /dev/tty simultaneously will not
read the data twice.

This change also modifies JsFile so it specifies whether the stream is
connected to a TTY when it is opened. This is also needed to support
piping data over stdin/stdout.

Change-Id: Icbf228e38622624137c2a80aa2bc124624863c6c
Reviewed-on: https://chromium-review.googlesource.com/256129
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Rob Ginda <rginda@chromium.org>
Reviewed-by: Dmitry Polukhin <dpolukhin@chromium.org>