Encode text to copy as UTF-8 before base-64 encoding
When copying text from Emacs in an hterm session, we must encode the text as
UTF-8 first. This is necessary to successfully copy non-ASCII characters.

Without this patch, you will see an error 'Multibyte character in data for
base64 encoding' in *Messages* if you try to copy any non-ASCII character.

Change-Id: I82ae3b4ba14646cea70fdddb2f77d6115cb95360
Reviewed-on: https://chromium-review.googlesource.com/275023
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Rob Ginda <rginda@chromium.org>