hterm: Fix mouse reporting
Previous code was busted because the terminal's 'reportMouseEvents_'
flag was initialized to off and not synchronized with the vt state
until a mouseup event.  This meant that mouse reporting didn't start
until after an initial click (and move, for some reason) of the mouse.

This patch changes 'reportMouseEvents_' to 'defeatMouseReports_',
which indicates only that the user is in the middle of a mouse gesture
started with alt-click.

Change-Id: I169e174f152a70d6236eb7a37dbe89dde332ec94
Reviewed-on: https://chromium-review.googlesource.com/278021
Reviewed-by: Marius Schilder <mschilder@chromium.org>
Reviewed-by: Renaud Paquay <rpaquay@chromium.org>
Tested-by: Rob Ginda <rginda@chromium.org>