hterm: set cursor as text for the whole space
Currently the cursor switches between text and pointer depending on
whether there is actually text under the cursor.  This can make the
cursor visually jump around when it moves between blank lines which
is jarring for the user.  By setting the cursor to text in the main
body element, it doesn't visually change in the whole port.

Change-Id: Iad57bfbd372689b6bdd5eeeb91040f3463ec87d2
Reviewed-on: https://chromium-review.googlesource.com/308630
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Rob Ginda <rginda@chromium.org>