ssh_client: improve errno handling when open fails
The open() wrapper sets errno to EACCES regardless of the actual failure.
With newer openssh versions, when it tries to read an ssh key that does
not exist, it ends up displaying error messages like:
	Load key /.ssh/id_dsa: Permission denied
And does so four times over since it tries to read that many types.

Plumb the PP error value back out of PathHandler and into the FileSystem
layer so it can translate it to an errno to pass it back to the caller.
Now the open() wrapper can correctly set the value to ENOENT which will
get ignored by openssh.

BUG=chromium:535763

Change-Id: I9fa83e825d7e114cb71ae1ef1b168d5d3e6bb759
Reviewed-on: https://chromium-review.googlesource.com/330931
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>