workaround race between bg page startup
When opening Chrome for the first time, sometimes the SecureShell window
can be loaded before the background page has finished being created.  As
it's only used to show update notifications (and for minor debugging),
let's just skip the binding if it doesn't exist yet.  Most people won't
notice either issue, and it'll be fixed the next time the app is run.

Change-Id: I6b3665f020ecc8ee3010e1bd0884c583998e0d3d
Reviewed-on: https://chromium-review.googlesource.com/337630
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>