hterm: disable ligatures
When using OpenType fonts, some include ligatures that mess with the
rendering (e.g. "ae" might be rendered as "æ" even though they are
really two bytes).  This messing with the display (the cursor can't
find the right location) and just confuses users in a terminal fixed
font setting.

Some notable examples of this appear to be recent OS X systems with
their Menlo font family.

This should not impact distinct Unicode codepoints.  i.e. U+00C6 is
æ and will continue to be rendered as a single character.  It should
only affect the (incorrect) rendering of two different codepoints as
a single one.

BUG=chromium:621519

Change-Id: I49f8afb950fa620fef87a8230a88a4f00e060889
Reviewed-on: https://chromium-review.googlesource.com/356530
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Rob Ginda <rginda@chromium.org>