nassh: crosh: add ctrl-shift-p hotkey for opening preferences
In CrOS, you can quickly load crosh with ctrl+alt+t, but you can't
easily get to the preferences screen for it.  People are forced to
find the options url by hand and load it in a tab.  Add a shortcut
so they can quickly load the options page.  We'll update the intro
help text to tell them about it to fix discoverability.

Change-Id: I1b8228e68834ca913c4e04dc7974dd370ab81de4
Reviewed-on: https://chromium-review.googlesource.com/389492
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>