libdot: fix two windows sharing prefs when resetting to defaults
If a window changes a preference to the default it will delete the
localStorage entry. When another window detects this event it should set
that preference to default. But currently it is set to the empty string.
This has caused errors when sharing display preferences between hterm
windows that share a profile.

BUG=chromium:685441

Change-Id: I424fa40f7bfdf90811bc16e4803c01ae5f8bf7f3
Reviewed-on: https://chromium-review.googlesource.com/429651
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>