nassh: send pasted input to the input buffer
After the tty rewrite, we can't send all input to fd 0 anymore as the
stdin and tty fds have been split.  When ssh asks for data on the tty
port, pasting it to fd 0 (stdin) means it gets queued.  Instead, send
the paste output to the new input stream so it gets routed properly.

BUG=chromium:705155
TEST=pasted & typed into password field and it worked
TEST=pasted & typed into ssh after connecting and it worked

Change-Id: I990abea3246970a077daf1235064a8a3742030a1
Reviewed-on: https://chromium-review.googlesource.com/461558
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>