nassh: register ourselves as ssh:// handler
Trying to register a protocol handler can only be done in response to a
user initiated event, so we can't register it automatically.  Instead,
we'll hook the connect button so we register as soon as someone connects
to a system.  The user will get a prompt at that time and can decline or
approve the registration.  It shouldn't show up again.

With multiple extensions, users can select their preference under the
standard handlers page: chrome://settings/handlers

Basically we add support for another command line form.  In-addition to
"nassh.html#profile-id:XXX", we support "nassh.html#uri:XXX".  Another
form is needed because the URI is encoded.  e.g. The full form is:
	nassh.html#uri:ssh%3A//vapier@vapier%3A22
We don't want to blindly unescape the command line as we might mangle
valid values otherwise.

We don't support profile names here since it doesn't make much sense:
profile names are inherently system-specific and <a> links in HTML pages
are inherently system-independent.

BUG=chromium:703777

Change-Id: Ic7d6c6d7d9d46ec353179416a37de8f73b5584fa
Reviewed-on: https://chromium-review.googlesource.com/461461
Reviewed-by: Brandon Gilmore <varz@google.com>
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>