ssh_client: strip executables
We build openssh with debug symbols (even for release) and then install
those.  These measure about 7.5MB.  If we strip all the debug info, it
cuts the size down to 2.5MB.  There's no need to ship the debug info in
the extension itself, so keep the debug info in a .dbg.nexe file.  It's
not like when there's a crash anyone actually debugs it directly.  We'll
still keep these in the bundled tarball we archive:
	gs://chromeos-localmirror/secureshell/releases/

Change-Id: Ibd917070aeabb52c336a117190d73462c6af2035
Reviewed-on: https://chromium-review.googlesource.com/464006
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>