nassh: integrate with omniboxes via the "ssh" keyword
This let's people type "ssh root@host" into the omnibox to launch a ssh
session in a tab.  They can use alt+enter to open in a window instead.

The downside is that it overrides people who select the omnibox and do
a search that starts with "ssh" -- it'll now go to this app.  You can
still do an explicit search by using "?ssh", or by hitting ctrl+k.  It
isn't possible currently to make omnibox integration an option as the
keyword setting is in the extension manifest.

BUG=chromium:711879

Change-Id: I8c8552186cb125e1532ed3906688d9e982cbaf8f
Reviewed-on: https://chromium-review.googlesource.com/479213
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>