libdot: update title during tests
For longer testsuites (like hterm) where the display doesn't seem to
change, it's not clear that tests are actually running.  Update the
window title before each test so it's clear we're making progress.
We also set the title when we finished all the tests.

To support this, we add preamble & postamble callbacks to the
TestManager class mirroring the TestManager.Suite callbacks.

We also fix some errors in the existing TestManager.Suite.postamble
logic: the comment had a copy & paste error, and the call forgot to
pass in the result/cx variables.

Change-Id: Ic276b95aa2dd8c58c80b1c98c591926d8721c02b
Reviewed-on: https://chromium-review.googlesource.com/483621
Reviewed-by: Rob Ginda <rginda@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>