hterm: support scrolling via touch events
This gets us basic functionality: we scroll based on touch events and
nothing else.  The more fingers you add, the faster you scroll.  We
might want to add more functionality in the future like long presses
start a selection, or maybe selections are controlled by one finger
and two fingers are used to scroll.  For now let's add this and see
how people like it.

BUG=chromium:221418

Change-Id: I69dc258a75ee4ad16d587ed2b65745e66d28aa36
Reviewed-on: https://chromium-review.googlesource.com/490748
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>