nassh: sftp: ignore command args from connection profile
If people have custom commands set up in the profile and they try to
connect via sftp, the connection setup will fail.  This can be a bit
confusing, and remote commands don't make sense to sftp, so ignore
the request.

There are other ssh command line options that don't make sense for
sftp, but we won't worry about those just yet as they aren't as easy
to handle as this.

Change-Id: Ica28acd3a0e2bca0b5b7b5da479a7e518632162a
Reported-by: Mike Maloney <maloney@google.com>
Suggested-by: Michael Samuel <michaelsamuel@google.com>
Reviewed-on: https://chromium-review.googlesource.com/490906
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>