nassh: improve invalid connection settings highlights
Instead of using the cursor color, always use red.  If the user has a
cursor color that is close to the background color (white), then you
end up not being able to read the invalid field and thus fix it (oops).

Instead of just coloring the text, put a red border around the field.
This makes it more clear which fields have problems or are missing.

Change-Id: I242fbd4d4cf47e3593c4ef011764a4c27844c478
Reviewed-on: https://chromium-review.googlesource.com/495771
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>