nassh: format preference JSON objects
Modifying some of the extended fields that take lists/objects can be a
hassle.  When the pref is saved, we run it through JSON.stringify, and
that strips out all whitespace giving us things like:
{"TERM":"xterm-256color","LANG":"en_US.UTF8"}
As you add more fields, it gets even worse.

Let's pass in whitespace to its spacer argument so we now get:
{
  "TERM": "xterm-256color",
  "LANG": "en_US.UTF8"
}

I went with two spaces instead of a tab as the latter is not easy to
type into the browser.

Change-Id: I769074dbc7e394d64fc7eee94a277350334cdab1
Reviewed-on: https://chromium-review.googlesource.com/497166
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>