libdot: allow i18n translations to refer to each other
Sometimes we want the same translation to be loaded into different node
attributes.  For example, the aria-label and title translations often
can re-use the same content.  This can be done today by doing:
  <input id='field-description' type='text'
         i18n='{"aria-label": "$id", "title": "FIELD_DESCRIPTION_ARIA_LABEL"}' required>

But this is unweildy when needed for more than just one field.  Instead,
let's add an "=" syntax so that one i18n attribute can refer to another.
Now the example above looks like:
  <input id='field-description' type='text'
         i18n='{"aria-label": "$id", "title": "=aria-label"}' required>

BUG=chromium:719070

Change-Id: I6ed43e35d84eed958405c19fe57b589192e86b8f
Reviewed-on: https://chromium-review.googlesource.com/502208
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Mike Frysinger <vapier@chromium.org>