hterm: add a sep option for pasting on right clicks
We currently only support pasting for one mouse button.  For most
systems, this isn't a big deal as users as used to only one button
pasting text (either because a middle button isn't common, or it
isn't commonly used).  But on X11/Linux and CrOS systems, we have
an intersection of users and mice where right & middle pasting is
expected -- Chromebooks only have two buttons, and you can only
right click w/the touchpad.   But we have devs coming from Linux
that use middle paste a lot, and they can hook up three button
mice to the system.

So let's add a dedicated option for pasting on right clicks, and
default it to on for all platforms.  This leaves the other paste
button to auto select the middle button, or for users to pick any
other button manually.

BUG=chromium:725703

Change-Id: I0150ede70d1371b8d7ab85837664b74a9f1317fb
Reviewed-on: https://chromium-review.googlesource.com/513446
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>