nassh/ssh_client: rework how we pass sftp subsystem down
Rather than re-use the -s argument to select the sftp subsystem, add a
new parameter to the plugin API so we can signal this independently of
the command line.  By itself, this isn't terribly useful (and is a bit
redundant), but in a follow up CL, it allows us to simplify the command
line parsing on the JS side significantly.

BUG=chromium:725292

Change-Id: I55d04617050a3f7c76e3ecf61138ef3fad2f711c
Reviewed-on: https://chromium-review.googlesource.com/586672
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>