nassh: rework user command line processing
Move the custom user command line options to the end so we always append
them.  This makes things "just work" more often when users specify remote
commands and behave much more like native `ssh`.  We also no longer need
to check the sftp mount state because the ssh plugin will throw away the
extra args for us.

We still don't handle quotes or embedded whitespace correctly, but we can
address that in a follow up if/when we supported quoting at all.

BUG=chromium:725292

Change-Id: Ib0d438cbc7a77366881cc7a058fc3d060f0085a1
Reviewed-on: https://chromium-review.googlesource.com/600637
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>