nassh: Implement an extensible SSH agent
The Agent class takes requests from the SSH client and passes them on to
a set of dynamically registered backends. The responses from the
backends are aggregated and reported back to the client. In the
connection dialog, the user can choose the backends used by Agent by
providing a list of backend IDs separated by commas as the value of the
--ssh-agent relay option.

So far, only the requests SSH_AGENTC_REQUEST_IDENTITIES and
SSH_AGENTC_SIGN_REQUEST are supported, together with their respective
responses.

This change is backwards-compatible: If the value of --ssh-agent is just
an extension ID, the old SSHAgentRelay is used instead of the new
SSHAgent.

The implementation introduces the following new classes:

 * nassh.agent.Agent: Initializes the backends, relays the requests to
    them and bundles up the responses.
 * nassh.agent.Agent.UserIO: Provides backends with uniform access to
    terminal IO.
 * nassh.agent.Backend: A minimal implementation of a backend which all
    other backends should derive from. It provides a backend under the
    ID 'stub' which can be used for testing purposes.
 * nassh.agent.Message: Parses and writes messages used in the SSH agent
    protocol.
 * nassh.Stream.SSHAgent: Inherits from nassh.Stream and relays raw SSH
    agent requests to the Agent class.

BUG=chromium:712699
Change-Id: I0a4dfc6681839518f98d9ed461479fd5a461d6bd
Reviewed-on: https://chromium-review.googlesource.com/550115
Tested-by: Fabian Henneke <fabian.henneke@gmail.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>