hterm: character maps: rename "name" to "description"
There are two different "names" we run into here: there's the name used
to uniquely identify the map (e.g. "B" for the US map), and then there's
the description text we provide for debugging (e.g. "us" for the US map).
Lets rename the debugging one to "description" to disambiguate.

BUG=chromium:734318

Change-Id: I051847ad3bc8c346d168cecb80ce31bb0c04a389
Reviewed-on: https://chromium-review.googlesource.com/539173
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>