nassh: Add an agent backend for smart card keys
Implements a 'gsc' backend for the SSH agent introduced in the previous
commit. This backend enables the use of private keys stored on
OpenPGP-enabled smart cards for SSH authentication. It relies on the
Google Smart Card Connector client library to communicate with the
Smart Card Connector app (khpfeaanjngmcnplbdlpegiifgpfgdco), which
offers an NaCl port of PCSC-Lite and works with many popular smart
cards.

So far, only the requests SSH_AGENTC_REQUEST_IDENTITIES and
SSH_AGENTC_SIGN_REQUEST are supported, together with their respective
responses. When asked for SSH identities, the backend will read the
authentication subkeys from all connected smart card readers.

The backend only supports the OpenPGP applet and thus only identities of
'ssh-rsa' type. Support for the PIV applet and ECC identities can be
added with only minor changes to the overall structure.

The implementation introduces the following new classes:

* nassh.agent.backends.GSC: Implementation of the interface
   nassh.agent.Backend which provides SSH identities stored on smart
   cards.
* nassh.agent.backends.GSC.CommandAPDU: Represents command APDUs used
   in smart card communication.
* nassh.agent.backends.GSC.DataObject: Represents a data object stored
   on a smart card.
* nassh.agent.backends.GSC.SmartCardManager: Manages the lifecycle of a
   connection to a smart card reader and provides convenience functions
   for the command APDUs used in the SSH authentication scheme.
* nassh.agent.backends.GSC.StatusBytes: Represents status bytes
   returned by smart cards.

BUG=chromium:712699
Change-Id: I1a650141a7047e7973d4702df185ea5f55592300
Reviewed-on: https://chromium-review.googlesource.com/550116
Tested-by: Fabian Henneke <fabian.henneke@gmail.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>