nassh: rework release note display conditions
Instead of basing it on version, base it on the contents of the release
notes itself.  This way we can keep adding notes in the dev channel and
they'll show up whenever we update.  Currently, users see it once, and
then not again since the "last release" base is the major stable ver.

We'll cheat and use the size of the notes as a poor man's hash.  It's
unlikely that we'll run into collisions here between two adjacent sets
of notes (since we only remember one value, and we reset it whenever
it changes).

BUG=chromium:718723

Change-Id: If88054603d6283d2c20d3847ee29bb737d83b03d
Reviewed-on: https://chromium-review.googlesource.com/575692
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>