hterm: default all character maps to US/ASCII
Currently we default G1 to the graphics map and G0/G2/G3 to US/ASCII.
This makes it very easy to switch to the graphics map (a single SO or
Ctrl+N), but it also means it's easier to accidentally select it when
you cat a binary file.

Since other terminals default all maps to US/ASCII by default (checked
eterm, iTerm2, konsole, mosh, libvte/gnome-terminal, rxvt/rxvt-unicode,
suckless, terminology, xterm), change hterm to do so as well.  The few
programs that use the graphics map will manually load it anyways since
they have no idea what any of the maps are currently pointing to.

BUG=chromium:735748

Change-Id: Iface24830f236a9c33bf54f9406ea2635bde2334
Reviewed-on: https://chromium-review.googlesource.com/582077
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>