libdot: rewrite lib.TestManager.Log
Rewrite this class for a few reasons:
- Provide an API that matches the console object so it's more natural
  to use and provides a fuller API (vs just "print")
- Use console grouping so we get real nested logging automatically
  (instead of the hand indented logic we have today)
- Easily capture all the output sent to the console so we can display
  it in the summary page

As a result, we purge the print()/flush()/pending_ logic entirely since
it was only used to support the ad-hoc grouping.  We get to classify all
the println users into proper info/warning/error categories which show
up clearly (and colorized) in the log now.

Change-Id: If0bc96e40d9f57e1fec3413cee4440473c0c4de3
Reviewed-on: https://chromium-review.googlesource.com/592909
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>