nassh: support quotes with ssh command line
This allows for basic quote support.  It works if the argument itself
is entirely quoted, but does not work with nested quotes.  If we want
to go that route, we'll have to rewrite significantly (see the bug
below for indepth discussion).  Hopefully this simple form will work
for most people's needs though.

BUG=chromium:725625

Change-Id: Ic570ea9a3ff497f4dbf7632ea3530baa5e78c5e1
Reviewed-on: https://chromium-review.googlesource.com/612042
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>