hterm: switch cursor positioning to use CSS vars
Calculating the position of the cursor in pixels is done entirely in JS.
The actual display of rows and their cumulative height is done in CSS.
This leads to a situation where, on some zoom settings, the precision of
floating point math in the two systems are slightly different leading to
skew in positioning.

Since we're already exporting the character size to CSS variables, we can
export the position (in cols/rows) to CSS too, and then let the CSS side
do the final calculation.  This simplifies the JS code, makes the CSS code
more readable, and improves the cursor positioning over zoom levels.

Change-Id: Ibcb98c8f37cede5f2538892a4eb4d260602498d0
Reviewed-on: https://chromium-review.googlesource.com/602793
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>