hterm: differentiate between non-ASCII and wide-characters
We need to treat strings specially when they're non-ASCII and not
just when wide characters.  i.e. When we're dealing with combining
codepoints.  Add an ASCII node property to the text nodes so we
can differentiate the states.

This doesn't fully fix combining characters, but should at least
improve things.  For example, they're still broken with wide chars.

BUG=chromium:654839

Change-Id: I4caeeadda4da019c557ea51e7e9fff4ec6634ef9
Reviewed-on: https://chromium-review.googlesource.com/615984
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>