hterm: split strings based on graphemes
The current logic tests strings on a codepoint-by-codepoint basis by
looking at their terminal cell width.  This mishandles graphemes that
are composed of multiple codepoints by breaking them up.

This can be easily seen with separate Hangul Jamos.  These three:
  ᄋ U+110B + ᅰ U+1170 + ᆫ U+11AB
should be rendered as a single glyph:
  웬 U+C6EC
But instead, they're broken up and are unreadable.

By switching to the Intl.Segmenter API, we're able to segment text
based on graphemes and these now render properly.  Note that that
API is based on UAX25 which mishandles some scripts (see bug 654839
for a bunch of examples), but we don't have much choice at this
point until the ICU/CDLR/Unicode world catches up.  We also won't
be that much worse than the status quo.

Performance wise, we take a hit of about 15% when processing non-ASCII
data which should be acceptable considering the increase in correctness.

Bug: chromium:654839
Bug: chromium:737454
Change-Id: Ifb129f7a820d0c3b9932917c9d046a3fa5404713
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/1460711
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>