libdot: wc: fix substring handling of surrogate pairs
The lib.wc.strWidth function was operating on codeunits (charCodeAt)
instead of codepoints (codePointAt) causing it to incorrectly count
and split surrogate pairs.  This doesn't normally matter as JS uses
UTF-16 for strings which means the two are the same for codepoints
in U+FFFF and lower (which are the most common), but when processing
codepoints higher than that (like U+2099d 𠦝 or U+1f60e 😎), we would
chop off a following character one too earlier.

Rewrite the function to operate on codepoints like we already do in
lib.wc.strWidth.

BUG=chromium:757476

Change-Id: I5891f56bcf1b26bc5639ee88c25de8560c1fd0cf
Reviewed-on: https://chromium-review.googlesource.com/624195
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Brandon Gilmore <varz@google.com>