hterm: fix OSC termination edge cases
Our OSC sequences may be terminated by either \a or a string terminator
(ST) \e\\.  However, the current logic requires that the ST be in the
same buffer.  If the lower layers happen to split the buffer such that
the first buffer ends with the \e and the next buffer starts with a \\,
we end up rejecting it because we saw the plain \e.  Normally this does
not happen, but with a large series of OSC sequences, it's easy to get
ssh buffering to split it up on us.

While sorting this out, we find the current logic also skips over any
embedded escapes in a buffer if there is a valid ST sequence in there.
We should be rejecting those too.

While debugging this, we notice that the timeout logic does not work
correctly either.  Due to order of operations, the string/date/math
creates a NaN which fails the abortReason test since NaN is falsey.

BUG=chromium:212700

Change-Id: I1c94a300f4afaf2c62fa122d8365d1a4dc209229
Reviewed-on: https://chromium-review.googlesource.com/627842
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>