crosh: always decode pasted strings
If you try to paste Unicode into crosh, the shell doesn't see a UTF-8
encoded stream.  This is because we never decode the stream when it
comes from a paste event, only when it comes from the keyboard, but
the terminal paste logic also encodes the content.  Change the logic
so we don't differentiate between keyboard & paste sources.

BUG=chromium:560635

Change-Id: I3af21f164ad37d403d0d5eff1fc36001292a24a1
Reviewed-on: https://chromium-review.googlesource.com/636307
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Brandon Gilmore <varz@google.com>