hterm: fix container matching logic wrt ascii & blink settings
The default state for the blink attribute is undefined, so when we
try to compare them against true/false values in the current state,
the test always fails.  Normalize the values before testing them.

The asciiNode test had a typo where it'd always check itself.

Further, always propagate the asciiNode property since the default
would otherwise be false when we really want it to be true.

With all these lessons in hand, write some tests, and sprinkle some
comments around the code to give guidance to people adding custom
attributes in the future.

Change-Id: I318c1ebd762e844256394be16689b563bc366ac5
Reviewed-on: https://chromium-review.googlesource.com/644532
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>