libdot: stack: rewrite and add tests
Rewrite to use (new Error()) directly instead of try{throw new Error}
then catch{ex}.  This makes the code simpler, and speeds things up a
little bit -- less than one millisecond, but when called thousands of
times, this can add up.

Also add a count option to limit the number of frames we extract.
This is useful when we just want the immediate caller.

Change-Id: Ic0200045f63c7511268525a9316373a83e4d7630
Reviewed-on: https://chromium-review.googlesource.com/1137974
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>