hterm: handle drag & drop nicely
There is a use case for this: if you're reading a document and you want
to copy over a paragraph, you could manually use copy & paste keys, or
you could drag & drop it into the terminal.  We also can smooth over
pasting of links to resources (like images), especially when the page
in question makes it difficult to copy & paste.

This also opens up an opportunity to let the user paste in the actual
source format (e.g. HTML) in addition to the plain text rendering.  We
use the shift key to differentiate the user intent.

BUG=chromium:402972

Change-Id: I0c97c65704c414fa715bf7bd5a068f73138605ac
Reviewed-on: https://chromium-review.googlesource.com/667363
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Brandon Gilmore <varz@google.com>