nassh: fix relay loop detection
The assumption was that session storage was scoped to a tab and thus
would not persist across connections (as that's what the comments in
our code around promptOnReload implies).  That's not the case -- the
active session storage is used to seed things when a new window is
opened via Ctrl+Shift+N.  That means the loop counter is not cleared
when we try to create another connection, or when we reconnect, and
ends up triggering a false loop error if you run a bunch of windows.

Clear the relay loop count at two locations where things should be
sane: when we display the connection dialog, and when we finish the
relay redirection.

Change-Id: Ia2822f3c02bfc13b9aabef4bab537c18aeabb473
Reviewed-on: https://chromium-review.googlesource.com/711715
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>