nassh: omnibox: allow matching saved profile names
The current omnibox supports connecting via a profile-id or a URI.
Feedback from users is that they want to type in a profile name and
connect via that.  We avoided that previously in case the profile
name had the same form as a URI (e.g. "root@foo").  If users do use
such forms, they can rename the profile as they like.

Change-Id: Id3fa2d81c12e76e46161882a0632136462fd68e3
Reviewed-on: https://chromium-review.googlesource.com/718176
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>