ssh_client: fix /dev/null reads
Reads on the /dev/null file should always return 0 and leave the buffer
alone.  Our NaCl shim was instead zeroing out the buffer and saying it
read the requested number of bytes.  Any code that tried to read the file
would get stuck in an infinite loop as the file never "ended".  We didn't
notice as this isn't a common scenario, but if you were to use it for an
option like UserKnownHostsFile, the ssh session would hang.

BUG=chromium:734431

Change-Id: Ia770faa475b2e6428eb1e6c98a44bb751c07a4be
Reviewed-on: https://chromium-review.googlesource.com/729526
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>