ssh_client: rework strtol usage to avoid lint warnings
The linter doesn't like us using strtol, but it suggests we use funcs
that are specific to the Chrome runtime instead.  Add a NOLINT flag to
kill that off.

That said, the two points using strtol are doing the same thing, so
unify them so they have equiv coverage in terms of error checking.

Change-Id: Ifbd019a9a8c3379168c55d3996acc06a426747bf
Reviewed-on: https://chromium-review.googlesource.com/729528
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>