nassh: add option to select different ssh client plugin versions
Add support for shipping multiple versions of the pnacl plugin.  This
can be helpful for people when upgrading in case there is a regression
in the plugin.  For example, when doing a version upgrade in OpenSSH or
libraries it uses.  It's not intended to support multiple versions for
a long time, just long enough to iron out regressions.

It can also be used to ship new versions for longer testing purposes,
like a WebAssembly implementation.

Change-Id: I9b3f65da7a15bdc7c2b0644213fca065d60e22a6
Reviewed-on: https://chromium-review.googlesource.com/771010
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>