hterm: move vt reset calls into the terminal layer
Our RIS (full reset) code path was resetting the VT itself, then calling
the terminal reset which would turn around and reset the VT again.  So
drop the VT reset from the RIS handler.

Our DECSTR (soft reset) code path wouldn't reset the VT twice, but anyone
using the terminal.softReset API wouldn't have the VT reset.  So move the
VT reset from the VT layer into the terminal layer.

Now the terminal reset API is consistent: users of the JS API will get
the right behavior, and the VT callbacks will follow the same codepaths.

We move the VT reset calls in the terminal functions to the top just to
be consistent code-wise for the readers.  It shouldn't matter here when
the VT was reset relative to the other features since there isn't any
overlap.

BUG=chromium:765254

Change-Id: I17b8312f1382f455eeb5faa6f3f43ba7d906ddac
Reviewed-on: https://chromium-review.googlesource.com/797992
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>