hterm: move cursor save/restore state from per-terminal to per-screen
Currently when you save/restore the cursor state (a confusing phrase as
the set of attributes are a bit more than just the cursor), it's done on
a per-terminal basis.  The active screen (primary vs alternative) is not
taken into consideration.  This is unlike xterm (where the screen concept
comes from) which has separate cursor states for each screen.

Refactor the CursorState class so it's no longer tracked in the VT layers
directly, but is part of the screen class, and the terminal class helps
mediate access to the active one (which is what mediates the screens and
associated content now too).

BUG=chromium:765254

Change-Id: I8043edc95c0bc7670ec480a39c64cf742f357398
Reviewed-on: https://chromium-review.googlesource.com/797994
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>