libdot: run replacements on default messages too
The current message manager helper will use the default text without
running any replacements on it.  This makes it difficult to process
arguments with the default text and actual localized messages.

We also tweak the get behavior so we automatically register the text
using the fallback value so we avoid warnings in the future.  Code
shouldn't be using different default values for the same id.

BUG=chromium:717743

Change-Id: I2d37879da4032756214aa6e0131b2e638d1fbc62
Reviewed-on: https://chromium-review.googlesource.com/810464
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Brandon Gilmore <varz@google.com>