hterm: drop max-string-sequence support
We have two ways of breaking out of escape sequences: either we hit a
byte limit (default of 100k), or we timeout (default of 20s).  The byte
limit was added before the timeout was, and is largely redundant with
the timeout logic.  If someone were to set this field too low (say to
0), things would misbehave in very weird ways (and has happened to at
least one person who didn't change it on purpose).

We now have a desire to increase the byte limit, perhaps to megabytes,
as we add support for things like displaying images inline.  So lets
drop the byte limit entirely.  If we did hit an unterminated escape
sequence, we still have the timeout to fallback to.

Change-Id: I777b0bd0841bad612a5ef3a7edaf404f93a7dc9d
Reviewed-on: https://chromium-review.googlesource.com/815620
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>