hterm: allow keybindings for unknown keys
If hterm sees an unknown key press, it will warn, but then stop all
processing.  This makes it impossible to add custom bindings in the
user preferences, or via the JS API if the key hasn't been defined
explicitly.

Lets change the logic so that we still warn, but then we automatically
register the key using the "UNKNOWN" key definition which defaults to
PASS behavior.  If there are any keybindings active, they'll get picked
up automatically, otherwise we get the same behavior as today.

Change-Id: If18f693c873c40401b9c46b18a3729e5b05cb10f
Reviewed-on: https://chromium-review.googlesource.com/822691
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>