nassh: mark sftp mounts as non-persistent across logouts
Since the current sftp code has no way of resuming mounts, mark it as
disabled using the new "persistent" field in Chrome 64.  Maybe some day
we'll have a way of trying to auto-resume (either using saved keys or
autoprompting for credentials), but until then, the user doesn't have
to go through every mount and explicitly unmount it on login.

Change-Id: If77415d79f92139fe277ad5cfb510440b63e3ebe
Reviewed-on: https://chromium-review.googlesource.com/822776
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>