hterm: fix typo in mouse wheel reporting
The comment says "add 64", but the code itself is adding "96".  The
xterm protocol says it should indeed be 64, and some ad-hoc testing
shows that xterm/libvte are reporting 64.  The original commit that
added this code used 96, but there doesn't seem to be any mention in
the commit message, Gerrit review, or bug report as to why.  Change
it to match xterm as that's what programs expect (e.g. tmux).

Change-Id: I24afa54eda55891fe21b2405dcc8d1dac065cb18
Reviewed-on: https://chromium-review.googlesource.com/826008
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>