nassh: fix resize errors after connect
After the connection dialog is closed, resizing the window will trigger
errors like:
Uncaught TypeError: Cannot read property 'document' of null
    at hterm.Frame.resize (nassh_command_instance.js:342)

This is because the resize handler we registered to move the iframe
around was left in place after we killed the iframe.  Add a cleanup
hook to remove that handler automatically.

Change-Id: I991c142a949cf203881791dc11a0cbbe7811a882
Reviewed-on: https://chromium-review.googlesource.com/844014
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>