nassh: Implement a PIN cache for GSC agent
Introduces a session-local smart card PIN cache. The user is prompted
to confirm usage of the cache during the initial connection process. If
the cache is used, further requests for the PIN for the same reader/key
ID combination are served from the cache without further interaction.

The cache implements a map from strings to Uint8Arrays and the data is
encrypted using a random WebCrypto encryption key that is marked
non-extractable.

Change-Id: Ic2e8d51ee1341e0b0979dec90bbf3eadb71787a8
Reviewed-on: https://chromium-review.googlesource.com/845581
Tested-by: Fabian Henneke <fabian.henneke@gmail.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>