hterm: send a signal when the frame is made visible
Newer Chrome versions have changed how hidden iframes are handled.
We can't focus the frame ahead of time (or its children), and we
can't calculate font metrics.  Update hterm to send a "visible"
message when we show the frame so the code can react for any last
tweaks it might need to make.

Change-Id: I68382cd47066c7c9dc6d0fd3c6785e900e94ba9d
Reviewed-on: https://chromium-review.googlesource.com/886242
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>