libdot: improve handling of escapes with string embedding
The current logic does not take into account any embedded escapes which
can easily break things.  For example, it might turn '...\x1b[...' into:
  '...\x1' +
  'b[...' +
Which is invalid in JavaScript.

Change the logic to not wrap lines at all if it has an embedded escape
sequence, and delay the single quote escaping until after we break up
the line.  This latter part might make the line "too" long, but since
we're already admitting "best effort" here for style, it's not that big
of a deal.

Also document that we strip out newlines.  This isn't new behavior, it
just wasn't explicitly documented previously.

Finally while we're here, clean up the end of the string a bit so we
don't have to end every embedded resource with a dummy ''.

Change-Id: I80450959188573d0a608298982cb42b5caf34341
Reviewed-on: https://chromium-review.googlesource.com/889938
Reviewed-by: Brandon Gilmore <varz@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>