hterm: update Chromebook media bindings
Now that the clashing Firefox bindings have been isolated, we can
add bindings for the remaining Chrome OS mediakeys and flesh out
the top row.  Especially now that the volume keys can be captured
and processed.

It seems we can see the power button too now, but we just pass it
as we can't stop it.

On the Pixelbook (eve), the media keys are different: the forward
button is missing, and there's a new play/pause button.  So a few
of the keys are mapped incorrectly.  Fixing this will require more
thought.

We capture and eat the assistant key by default.  It doesn't seem
to be that useful for devs at the terminal, and they can always
add to their keybindings {"153":"PASS"} to keep it.  If it makes
more people unhappy, we can see about adding a dialog/preference.

BUG=chromium:807513

Change-Id: Ia97349b51a4f699c8bcfdaf497ba764483106ef3
Reviewed-on: https://chromium-review.googlesource.com/899455
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>