nassh: improve gnubby extension probing
Users are reporting SFTP mounting sometimes not working.  Suspect the
reason is because, on their CrOS systems, they only have the gnubby
app installed, and we end up defaulting to the extension.

The probe code has a timeout of 1 second, and when you first log in
to CrOS, there's a bit of a "thundering herd" problem in that a lot
of things are running and overall it's slow.  That causes us to stop
waiting for any results and go with the default (the extension).
This is all done in the context of the background page which may be
further deprioritized.

Making normal SSH connections works because it sets up a new runtime
everytime the window is loaded which does a new probe and doesn't run
into the timeout.  When making a SFTP mount, the background page does
everything, so the first probe at login has stuck (i.e. the one that
hit the timeout and defaulted to the extension id).

We address this in two ways: first we tweak the default so we use the
stable app id on CrOS and the stable extension id everywhere else.
This should make things "just work" for the majority of users.  Second,
we remove the overall timeout from the probes because it isn't actually
necessary: the probing is done asynchronously via Promises, so we don't
block any execution, and if nothing is installed, we simply don't pass
back any values.

We also add the dev variants of the extension in case the user has those
enabled instead of the stable one.

BUG=b:66217265

Change-Id: Id4d4aec20befc0a0f844b1f7a82c4de55718958c
Reviewed-on: https://chromium-review.googlesource.com/1081758
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>