ssh_client: drop support for non-PNaCl builds
We haven't used the NaCl/glibc code paths in a long time, and it's
extremely unlikely we'll ever care again, so purge them to clean up
various cruft.

PNaCl only supports newlib, so we can inline that assumption too.
That allows us to drop getdents/dirent logic which is unused, as
well as glibc-specific func wrapping hacks.

Change-Id: If421dc2f263f8b3cf99a599865ea9db4730b03be
Reviewed-on: https://chromium-review.googlesource.com/1205833
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>