nassh: Add support for RSA and ECDSA via PIV applet
Adds support for SSH authentication using RSA or the NIST P-256/P-384
curves on a PIV-enabled smart card to the 'gsc' agent backend.

The agent loops over all supported applets on all connected smart cards
and aggregates the SSH identities, which are then advertised to the SSH
server. When the user is prompted for a PIN, the applet is now
mentioned as part of the prompt.

We slightly deviate from the PIV standard in that we also support
authentication using the P-384 curve. Curve detection relies on the
algorithm parameters of the authentication certificate to contain the
curve OID, by default P-256 is used, which aligns with the spec.

The implementation relies on the external libraries pkijs and asn1js,
which are compiled into a dependency file using a rollup shim.

Change-Id: I0efed7d89615f6a3bc75733d956f4a33d228c564
Reviewed-on: https://chromium-review.googlesource.com/c/apps/libapps/+/757503
Tested-by: Fabian Henneke <fabian.henneke@gmail.com>
Reviewed-by: Mike Frysinger <vapier@chromium.org>