hterm: improve & document mouse cursor customization
When using url() with mouse cursors, the fallback keyword is required.
Rather than force everyone to use ",text" or ",auto" when customizing,
add it to the default rule.  If they did want to set a fallback value,
they still can as multiple ones are allowed.

While cleaning that up, make sure our rule for pointing at URL links
respects the user's customization.

Then document the whole process in the FAQ.

Finally, tweak the style priority so that the user's style choices are
higher than that of the stock ones.  Previously, the user's styles were
inserted first, then the stock one.  If they wanted to change stuff,
they had to use !important in their rules.  But even this doesn't work
for CSS variables which meant they couldn't change them.

Change-Id: Ie4c5df3c7e9dfacfc65441a9b30900cd57f67cc3
Reviewed-on: https://chromium-review.googlesource.com/1112052
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>