hterm/nassh: Make it possible to enable/disable accessibility at runtime
Currently we will decide whether to turn on some accessibility features
in nassh and crosh at the startup of the terminal. This means that if
the screen reader is then subsequently enabled/disabled we won't respect
the user's setting and (for example) some accessibility features won't
work until the app is restarted. This change adds a listener to detect
when the setting changes and turn those features on or off accordingly.
Tests have also been added for these cases.

Bug: 822490, 646690
Change-Id: I3e91ed78b8c2c45339c029b6013553647956be82
Reviewed-on: https://chromium-review.googlesource.com/1086675
Tested-by: Raymes Khoury <raymes@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>