hterm: Announce messages shown by showOverlay
This adds accessibiliy announcements for messages shown by
Terminal.showOverlay, for example the announcement for copying text.

A test is added. Relatedly, these announcements should be voiced even
when full accessibility support isn't enabled as there is no performance
cost to doing so. This removes the check which prevented those
announcements being voiced.

Bug: 822490, 646690
Change-Id: I6999f2c811534f081c465d93b3ee7baf971633c4
Reviewed-on: https://chromium-review.googlesource.com/1117729
Tested-by: Raymes Khoury <raymes@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>