hterm: Simplify handling of duplicate live output
This changes how duplicate live output is handled. Previously we would
wait for a delay, clear the live region, wait for a delay again and then
update it with the new text to output. This is necessary so that the
cleared live region registers as an attribute change with the screen
reader and then the duplicate text is also registered as an attribute
change.

Instead of doing this we can simply check if the previous text in the
live region matches the current text we're about to display. If so we
slightly change the new text so that an attribute change will be
registered.

This approach is also added for the assertive live region so that it
will properly handle duplicates properly. Tests are.

Bug: 822490, 646690
Change-Id: I1a75c55315abd14314960248ad0c695ccb9102c1
Reviewed-on: https://chromium-review.googlesource.com/1117726
Tested-by: Raymes Khoury <raymes@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>