hterm: Handle announcements of selection changes
We currently don't make any announcements for text that isn't contained
in the terminal output. For example if a character is deleted or the
user moves the cursor over previously edited text. This CL adds code to
announce these kinds of changes. This is done by listening for cursor
changes and heuristically trying to detect what kind of a change
happened based on the location of the cursor, changes to the row of
text that the cursor is on and whether a key event has been recently
received.

Note that a heuristic is the best that can be done here because we
can't distinguish input from output in terminal and so can't know which
changes to the cursor/terminal are caused by the user or the
application. Also, some announcements would be inappropriate in certain
terminal applications/contexts. A best-effort approach must be taken
where we handle the common cases.

Change-Id: Ic71df5bcc67583caecd163620e287eb1dfa29ce8
Bug: 822490, 646690
Reviewed-on: https://chromium-review.googlesource.com/1121901
Tested-by: Raymes Khoury <raymes@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>