libdot: i18n: move replaceReferences helper here
This is a standalone function and makes more sense in this dedicated
i18n module, and we'll use it more in the future here.

We also take the opportunity to fix/improve it and add some tests.
Specifically, we now allow the substitutions to be empty, missing, or
not have enough elements.  This brings the behavior in line with how
Chrome replaces things, and lets us simplify getMessage a bit.

BUG=chromium:857538

Change-Id: Iecc54f7d17c9ab773e64a7fc16fb4151219de6f5
Reviewed-on: https://chromium-review.googlesource.com/1135780
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Raymes Khoury <raymes@chromium.org>