nassh: make connectToArgString fully async
The connectToArgString function is a mix of async and sync exit logic.
For errors that be diagnosed immediately, we return false directly and
let the caller exit.  For all other errors, the failing code will exit
itself.  This causes a mix of UI error paths for the user.

It can call 4 funcs: mountProfile, connectToProfile, mountDestination,
and connectToDestination.  The first two always return true.  The other
two return false only if the destination fails to parse.  Once we push
that check down, they too always return true.

Once we remove all the true return code paths, it allows us to simplify
the callers and make it so we get consistent behavior regardless of why
the sub-command failed.  It also gives us a bit more flexibility with
these functions to add more async behavior.  We aren't doing that here,
but we probably need to for some relay changes.

Bug: 776106
Change-Id: I9eece4b50fed76da1b3919b45b86b549131b9a65
Reviewed-on: https://chromium-review.googlesource.com/1137976
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>