hterm: Update the selection when accessibility is enabled even if it's collapsed
This updates the html cursor (selection) to be in the same location
as the terminal cursor. We generally only do this if the selection
is collapsed to prevent clearing a selection that a user has set.
However when a screen reader is enabled, it is expected that
input should cause the selection to move to the cursor. So in this
case we always update selection, even if it's collapsed.

Bug: 822490, 646690
Change-Id: I455b29e18dbdc0365baef0f14aad9400bd348c79
Reviewed-on: https://chromium-review.googlesource.com/1133604
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Raymes Khoury <raymes@chromium.org>