hterm: Ensure the Page Up button isn't shown when the ScrollPort is focussed
Currently when the ScrollPort is focussed (e.g. when pasting data from
the clipboard) it will cause the PageUp button to be shown. This is
because selection is set to the start of the <x-screen> element. Instead
we set selection to the caret position, or if the caret is offscreen to
the last row in the terminal. This also improves things from an
accessibility standpoint.

Bug: 863668
Change-Id: Id87335b63a4e799e3c2ace4012cb6c3b7a97e707
Reviewed-on: https://chromium-review.googlesource.com/1137965
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Raymes Khoury <raymes@chromium.org>