hterm: turn default pref settings into proper objects
Working with the hterm.PreferenceManager.defaultPreferences object is
a bit of a pain: each entry is an array and people have to hardcode
the indices when accessing each field (and remember what each number
means).  Trying to add new fields is painful as we have to reindex.
Using a real object allows us to name the different fields.

The churn here is reindenting the arguments, and while we're here,
convert the help text to use `...` instead of '...' so we don't have
to escape embedded single quotes.  The content otherwise shouldn't be
changed.

Bug: chromium:717742
Change-Id: If04b56ce82f83c8469142145477387a59534c630
Reviewed-on: https://chromium-review.googlesource.com/c/1218682
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>