nassh: restrict the automatic description expansion
The automatic description->{user,host,port} expansion in the conneciton
dialog is a bit too greedy: it grabs common characters that are invalid
in hostnames like whitespace.  Simplify the regex a bit so we don't try
to grab everything in there.

The hostname match is still not perfect, but hopefully this should be
better for most users.

Change-Id: I64e0f6f8cbfdecdb75f5464ec2c08f56d9d86da9
Reviewed-on: https://chromium-review.googlesource.com/1236518
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>