nassh: rework our datetime stamp in the generated version
While Chrome is OK with loading local extensions with versions like
"0.9.20180928.145140", the CWS rejects it.  It enforces the manifest
version policy where each dotted field is in the range [0..65535].

Rework the generated datetime stamp to pack the date/time components
into the 16bits each one is limited to.  If we had 32bits to work
with, we could have used bit shifts to be a bit cleaner.

Change-Id: I3eab0153535a1c61c9802d086345e9d5a0fd79c8
Reviewed-on: https://chromium-review.googlesource.com/c/1252402
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>