nassh: mkzip: revise datetime stamp (again)
The algorithm I had in mind when I wrote this before had properties
related to scaling that I didn't preserve, so it ends up failing.
Basic math operations shows that we were simply adding the different
fields together which violated the monotonic increasing requirement.
  ($1 + 12 + 31) + ($2 + 31) + $3  ->
  $1 + 12 + 31 + $2 + 31 + $3 ->
  $1 + $2 + $3 + 12 + 31 + 31 -> ($1 + $2 + $3) + 74

Rewrite it again so that each field is scaled properly.  For the
month/days, we switch to the day-of-the-year field so we don't have
to pack it down ourselves.  For the hour/minute/seconds, we have to
scale down the seconds (by half) in order to fit into 16bits.

Change-Id: I09a6420374f55ea3a3caa18d5e514bcccf57ae22
Reviewed-on: https://chromium-review.googlesource.com/c/1264995
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>