libdot: mkzip: improve manifest handling
The current manifest rewriting code uses awk and makes assumptions
about the style/format.  Use some inline python to avoid that and
to simplify the code a bit.

We also delete the "platforms" key when --nouse_manifest_platforms
is passed.  The _platform_specific/ paths won't exist which makes
the CWS reject the extension if the key is in the manifest.

Change-Id: I243adcf485954a9a96748dd2f40f1b2e147bf069
Reviewed-on: https://chromium-review.googlesource.com/c/1267616
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>