nassh: sftp: improve server version packet parsing
The version packet from the server can be parsed like any other
response packet, other than the fact that requestId is not set.
Since we're already stuffing the pendingRequests_ with 'init' as
a fake requestId, lets use that in our version packets too.

This lets us drop the custom VERSION check in the onPacket call.

While we're here, lets parse out the optional extensions the
server might send over in its VERSION response.

Bug: chromium:722175
Change-Id: I7c20b49f3e37aaa00767ea68674fffb8dda0a429
Reviewed-on: https://chromium-review.googlesource.com/c/1286249
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>