ssh_client: openssh: simplify the build a bit
We currently put "-pnacl" in the libs we create.  I guess we did this
so we could keep it separate from "-nacl" in the output/ dir, but that
doesn't work well when we write all the objects from the plugin src/
w/out any suffixes to the output/ dir.

Since we no longer support NaCl (only PNaCl), and because we're going
to take a radically different approach for separating out wasm, drop
the "-pnacl" suffix everywhere.

While we're here, move the cache vars from the env to the configure
invocation.  This will help with later build migrations too.

Change-Id: I816d7c3ac649839068bf419e83efe9bff6a63e8f
Reviewed-on: https://chromium-review.googlesource.com/c/1289810
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>