ssh_client: ingest the nacl toolchain settings
The nacl-env.sh script from webports is a bit complicated and does
more than we need from it.  Add the few vars we care about to our
build script so we can drop it.

This will make it easier to refactor the code too as we add wasm
support and drop webports entirely.

Change-Id: Iaec8767f5fbf5eab7a4035020b867b933f48b8e7
Reviewed-on: https://chromium-review.googlesource.com/c/1289816
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>