nassh: sftp: use larger read/write sizes with OpenSSH
Since OpenSSH's sftp server uses 256KiB packet sizes (and always has),
it helps performance if we use a larger packet size.  We make this
decision based on one of the protocol extensions that OpenSSH has when
it starts up.  It's not fool proof, but let's see how it works in
practice in the real world.

Similarly for read packets, increase from 32KiB to 64KiB.

Change-Id: I75f5bea4c743afd947ab1f7d430176d8791f9a02
Reviewed-on: https://chromium-review.googlesource.com/c/1392320
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>