nassh: add an API for opening crosh windows
We let people access crosh.html directly from the web so they can pop
open windows (notably, the Crosh Window extension).  However, this lets
any website link to us which is undesirable, so extend the existing API
to include a 'crosh' request.  Only Chrome apps/extensions will be able
to access this endpoint.

This also allows the Crosh Window to probe multiple extensions in case
the user has the stable & dev versions installed simultaneously.

With this in place, we can drop crosh.html from web_accessible_resources
as it seems to have been added only for this purpose.

Change-Id: I1e75413c45c114bf693494d3d79012dd9b6a9f25
Reviewed-on: https://chromium-review.googlesource.com/c/1343301
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>