nassh: prefs: drop old relay-host/relay-port migration
This code has been broken since it was added in 2013 -- it has a typo
when it tries to set the "relay-option" field (it's "relay-options")
which causes the code to fail.  Anyone hitting this code path probably
reinstalled the app to clear their prefs and thereby clear these.  So
drop the unused migration code.

Change-Id: I891d05058e59e327428a4e6c0952e765bcac0e21
Reviewed-on: https://chromium-review.googlesource.com/c/1352190
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>