nassh: fix handling of default port settings
The default port value is null, not "".  When the port field is empty,
it will be an empty string which means we end up storing "" everywhere
instead of eliding it entirely.  Change it around to null instead.

This doesn't automatically clear out existing prefs, but it should get
cleared whenever the user displays/tweaks a specific connection.

Change-Id: I4e7fa0ccd1894df81f294b48fe4e054d44c23c84
Reviewed-on: https://chromium-review.googlesource.com/c/1352192
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>