nassh: rip option parsing out of nassh.GoogleRelay
This has been a long standing TODO -- the relay-options field started
out as just relay options, but then it grew more things unrelated to
relays, but we never ripped the logic out of the relay object.  Do it
now in preparation for adding even more logic unrelated to relays.

This largely just moves it out of nassh.GoogleRelay and to the common
nassh.CommandInstance code, but we also drop the "known options" check
and let the higher levels blindly tokenize everything.  This makes it
easier to pass relay-specific options down without having the higher
levels care.  We still do validation in the html code, so this ends up
being a bit redundant.

Bug: 776106
Change-Id: Ibb5c90b1bb13589d4a2d53f61b787dd168b26910
Reviewed-on: https://chromium-review.googlesource.com/c/1352932
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>