nassh: add SSH-FE relay support
This is a bit rough, but largely works.

The nassh.Agent logic needs reworking, but that module needs a lot of
rework to better use typed arrays, so we won't get into that here.

Documentation is missing, but we'll add that once things settle.

The ssh key selection relies upon Google-specific agent messages which
means it's limited to gnubbies currently which is unfortunate.

The ACK protocol is an artifact of the frontend proxies, and is the
same as implemented in nassh_stream_google_relay.js, although the
implementations are independent currently.  It'd be nice to unify
them somehow.

Along those lines, we don't support connection resumption.  But the
current SSH-FE servers don't either, so it's not like we've lost any
functionality.

Bug: 776106
Change-Id: Ifc74e3ce4f1d3636b55dfb9a7d2d8468cde20997
Reviewed-on: https://chromium-review.googlesource.com/c/1390226
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>