libdot: add a lib.f.openWindow helper for noopener
Make sure when we open new windows (especially links to other sites)
we clear the opener field.  We add a helper to deal with this logic
so we don't have to replicate it everywhere, and because we need to
use a more verbose form to work on all current browsers.

More details on noopener can be found here:
https://developers.google.com/web/tools/lighthouse/audits/noopener

Change-Id: I79ebbb2123969459e2129f4db1e5b14277308a75
Reviewed-on: https://chromium-review.googlesource.com/c/1357959
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>