ssh_client: migrate from jsoncpp to native ppapi vars
We use the jsoncpp library to marshal messages between the JS and the
plugin layers which means we only ever transfer a string.  In reality,
this is overkill for what we need and introduces overhead: we transfer
dictionaries, arrays, integers, and strings.  The ppapi code offers
native types for all of these things.  So we can easily let the code
pass objects directly back and forth and avoid converting between JSON
strings all the time.

This also sets us for further future optimizations, the latter which
requires we pass objects directly via ppapi code:
- Instead of converting data we read & write to base64 strings, we
  can pass an array of integers for the byte stream directly.
- Once we operate on arrays of bytes, we can switch from the high-
  level Array type to the lower level TypedArray.
- Once we have TypedArrays, we can use native Text{En,De}coder APIs
  instead of our custom lib.{en,de}codeUTF8.

We don't do any of those things yet, but it explains the background.

This also has the benefits of dropping a package entirely (jsoncpp)
which shaves off ~200KB or about 10% of the compiled plugin.

Change-Id: Icc819b63c2f38f9fd52993c6dce413c31a322374
Reviewed-on: https://chromium-review.googlesource.com/c/1358159
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>