hterm: Initialising iframe contents asynchronously on FF.
Hterm's ScrollPort uses an IFrame into which content is dyamically
inserted. FireFox overwrites any content inserted into IFrames this way
when it actually loads a blank src. This was previously avoided by
setting the IFrame src to "#", preventing any page load which would
overwrite the dynamic content. However, FireFox 64 changes this
behaviour and src="#" overwrites the content, breaking hterm in FF64.
In order to work on FireFox we need to wait until the blank src is
loaded into the IFrame and only then dynamically insert content. This
approach only works on FF.

Bug: b/121029330
Change-Id: I19f878f439d2ddd6b2c4294369734d6b0dcf4838
Reviewed-on: https://chromium-review.googlesource.com/c/1381751
Reviewed-by: Mike Frysinger <vapier@chromium.org>
Tested-by: Adrián Pérez-Orozco <adrianpo@google.com>