hterm/nassh: switch to array buffers for sending tty data
Now that the ppapi plugin can handle array buffers, we can use the
native TextEncoder API to convert the UTF-16 JavaScript strings into
the UTF-8 byte stream the plugin wants.

This allows us to stop creating a UTF-16 string of UTF-8 codeunits,
and to stop base64 encoding/decoding just to transfer the data.

The terminalPrivate API that crosh uses handles UTF-16 strings directly
so no changes are needed there -- we've been checking the encoding
option only because hterm was reencoding things on us.

This means nothing relies upon the send-encoding value anymore in
hterm.  Since we've been all in on UTF-8, lets just drop it and wait
to see if anyone notices or cares.

We do the nassh/crosh/hterm updates in one commit as it's not easy
to break these up without leaving things broken inbetween.  And if
we're just deleting all the code, it doesn't seem worth the effort.

Change-Id: Ic0e8ce3929350018caf229d2ef12899905497b4a
Reviewed-on: https://chromium-review.googlesource.com/c/1392316
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>