nassh: switch to array buffers for sending sftp packets
We already had helpers to create array buffers, so switch to that to
avoid the base64 conversion.  We should do better and not serialize
to strings internally at all, but that'll take a lot more rework.

This changes the toByteArray API from returning a Uint8Array view to
the underlying ArrayBuffer, but no one is using this helper, so it's
not a big deal to rework this.  We want to pass around array buffers
anyways and not typed array views.

Change-Id: If561f811ec97ef57309299f5b198e3de28029a81
Reviewed-on: https://chromium-review.googlesource.com/c/1392317
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>