hterm: push selection logic down into copySelectionToClipboard
In preparation for using newer permission & clipboard APIs (which are
all Promise based), rework this API to accept the content we want to
copy directly.  If we have to use the document.execCommand method to
copy the data, we'll create the required nodes on the fly.  This helps
break up the changes into separate commits: converting existing APIs
before supporting newer APIs.

This step is particularly important because the newer APIs are both
asynchronous and take the state to be copied directly.  The current
execCommand API operates on the implicit state of the document and
whatever selection it has active (which is why we had to generate the
random nodes in the first place).

This also allows for more directed unittest coverage.

Bug: chromium:758466
Change-Id: I9a384d790b2ef0d73d30819ab03a16bbedf2d656
Reviewed-on: https://chromium-review.googlesource.com/c/1393864
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>