nassh: switch to array buffers with ssh-fe@google.com
This avoids the string/base64 reencoding overhead, although we still
have a memcpy in place in order to skip the leading ACK bytes.  We
could improve the JS->NaCl API to include byte offsets & lengths, but
I'm a little afraid we'd get in the habit of copying large buffers
over just to scan out a few bytes, and the offsets would lull us into
a false sense that we're more efficient than we actually are.

Change-Id: Ic4d30810207532d657c4e6e1802a6e87ec3d8527
Reviewed-on: https://chromium-review.googlesource.com/c/1393873
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>