hterm: switch to ES6 Symbol
ES6 provides a Symbol object specifically for creating unique objects.
We can drop our custom lib.f.createEnum API in favor of it.

We need to rework the hterm.TextAttributes logic a bit as we were
relying on the implicit string value of DEFAULT_COLOR being '' when
comparing to style properties (which default to '').

Change-Id: I97cb4a8608bea3f647896ec67c20bc35003d2942
Reviewed-on: https://chromium-review.googlesource.com/c/1396051
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>