hterm: workaround touch focus bug on CrOS
On CrOS, touch events in a window will not raise that window to the
foreground.  Since hterm captures touch events to provide its own
scroll/gesture logic, this means touching the window when it's in
the background leaves it in the background even as we process the
events generated.

On Windows, Chrome will automatically raise the window for us.

So add a workaround for CrOS to force focus until Chrome is fixed.

Bug: chromium:917938
Change-Id: If06f197b74b1b6b9b6b382dd3f85408d57a33cf9
Reviewed-on: https://chromium-review.googlesource.com/c/1396046
Tested-by: Mike Frysinger <vapier@chromium.org>
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>