nassh: drop update available check
The update available check was more useful when Chrome's update logic
wasn't as smooth, Secure Shell's background page was long lived, and
the app was under more active development.

Chrome has since changed such that it will queue updates and switch
over to the new version seamlessly when existing instances finish.

Secure Shell's background page was changed to not be persistent once
Chrome supported the concept, so it's automatically killed when it
isn't being used.

At this point, Secure Shell is pretty stable, and our stable release
schedule has slowed to less than once a month.

Taken altogether, this update check is not useful, so punt it.  I
don't even remember the last time I saw it kick in -- it was years
ago.

Change-Id: I42d304ea4bd6f8d331726d86ab7edb67add695c6
Reviewed-on: https://chromium-review.googlesource.com/c/1396048
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>