nassh: switch to chrome.runtime.getBackgroundPage
The chrome.extension.getBackgroundPage is synchronous and won't create
the background page if it doesn't currently exist.  We need to use the
chrome.runtime.getBackgroundPage version which is asynchronous as uses
a callback as it'll create the background page on demand for us.

This mostly worked before commit c4b90ef4973513b8e9052f0cff56e8717dc9f
("nassh: drop update available check") because that called the runtime
variant for us (which created the page), but even then that wasn't a
bulletproof setup.  If someone opened Secure Shell and let it idle,
and the background page timed out and was automatically destroyed by
Chrome, then we'd probably hit the same failure.  Code before that
commit also kept the background page reference in window.app_ which
might have changed the idle logic, but I don't know how reliable that
was, and we had some edge cases still.

At any rate, switch all the users of chrome.extension.getBackgroundPage
over to the runtime variant to avoid this non-determinism.

Change-Id: I5c16acb4eb34fc33a6310e160fadabbe98a6a03b
Reviewed-on: https://chromium-review.googlesource.com/c/1401206
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>