nassh: allow bookmarks to open as windows
Since Chrome always opens bookmarks as a tab and offers no option to
change this, we have to add code to handle it ourselves.  Add support
for a new 'openas=window' query string setting so Secure Shell will
automatically reopen itself as a window.

Since Chrome doesn't let tabs open windows with chrome=no, we have to
leverage the background page (which is in a more trusted context) to
make the window open call for us.

We also tweak the context New Window setting to do the same.  Before,
it'd open a new window but with the location bar, and not be able to
capture the Ctrl-N/etc... shortcuts.  For now we assume that anyone
who opens Secure Shell in a tab and then right clicks to open a new
window wants it to be a real window and not a looks-like-but-isn't-
really a standalone window.

Bug: 747003
Change-Id: I321e41125b1f22a0de5399c1d19773ac6a16e412
Reviewed-on: https://chromium-review.googlesource.com/c/1408090
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>