hterm: vt/io: move UTF-8 handling to hterm.Terminal.IO
The terminal interpret function uses a single stateful UTF-8 decoder
even when the io module is swapped.  This doesn't make sense -- the
byte stream is tied to a specific io instance, so lets rip it out of
the VT layer and move it to the IO layer where it makes sense.

This has the side-effect of changing the hterm.Terminal.interpret
API to accepting native JS UTF-16 strings instead.  This simplifies
things for callers as they no longer have to encode into UTF-8 from
other JS code when native JS strings are so much easier.

Change-Id: I2b6f7d82245bc2d8b5a687b30ecfbdd30ebc0561
Reviewed-on: https://chromium-review.googlesource.com/c/1438457
Reviewed-by: Vitaliy Shipitsyn <vsh@google.com>
Tested-by: Mike Frysinger <vapier@chromium.org>